requirements.txt
polars==1.7.1
pyarrow==17.0.0
numpy==2.1.1
scikit-learn==1.5.2
google-cloud-storage==2.18.2
pytest==8.3.3


